combine solutions in one script
The same array is passed as the "locs" queue and the "visited" set. This is only
safe in gawk, because "gawk sets up the list of elements to be iterated over
before the loop starts". Then the first "locs" queue is kept as it is also the
"visited" set and future "nextlocs" queues are discarded after scanning.